updated script

+++ shell_setup/bootstrap.py
@@ -13,10 +13,12 @@ __author__ = 'Gaurav Chauhan(gauravschauhan1@gmail.com)'
 DOTFILES_REPO_HTTP_URL = 'https://dotgc@github.com/dotgc/dotfiles.git'
 SPACEMACS_REPO_URL = 'https://github.com/syl20bnr/spacemacs'
 GIT_COMPLETION_URL = 'https://raw.githubusercontent.com/git/git/master/contrib/completion/git-prompt.sh'
+VUNDLE_REPO_URL = 'https://github.com/VundleVim/Vundle.vim.git'
 
 HOME_DIR = os.path.expanduser('~')
 FS_DOTFILES_PATH = os.path.join(HOME_DIR, 'dotfiles')
 EMACS_LIBS_DIR = os.path.join(HOME_DIR, '.emacs.d')
+VIM_DIRECTORY = os.path.join(HOME_DIR, '.vim')
 BACKUP_DIR = os.path.join(HOME_DIR, 'old-dotfiles.backup')
 
 DOTFILES_TO_IGNORE = frozenset(['.gitignore', '.git'])
@@ -34,7 +36,7 @@ def setup_dotfiles():
             target_symlink_path = os.path.join(HOME_DIR, file_name)
             if os.path.exists(target_symlink_path):
                 shutil.move(target_symlink_path, os.path.join(BACKUP_DIR, target_symlink_path))
-                time.sleep(200)
+                time.sleep(0.3)
                 '''
                 if os.path.islink(target_symlink_path):
                     os.unlink(target_symlink_path)
@@ -62,6 +64,14 @@ def setup_emacs():
     subprocess.call(['git', 'clone', SPACEMACS_REPO_URL, EMACS_LIBS_DIR])
     logging.info('Emacs setup done')
 
+def setup_vim():
+    logging.info('Setting up vim')
+    if os.path.exists(VIM_DIRECTORY):
+        shutil.rmtree(VIM_DIRECTORY)
+    subprocess.call(['git', 'clone', VUNDLE_REPO_URL, VIM_DIRECTORY + '/bundle/Vundle.vim'])
+    logging.info('Vim - vundle setup')
+
+
 def setup_tmux():
     pass
 
@@ -71,9 +81,15 @@ def setup_git():
 def download_dotfiles():
     logging.info('Downloading dotfile')
     if os.path.exists(FS_DOTFILES_PATH):
-        shutil.rmtree(FS_DOTFILES_PATH)
+        logging.info("dotfiles exists. re-download? (y/n) - default y")
+        ans = raw_input() or 'y'
+        if ans == 'y':
+            shutil.rmtree(FS_DOTFILES_PATH)
+        else:
+            logging.info('not downloading dotfiles again')
+            return
     subprocess.call(['git', 'clone', DOTFILES_REPO_HTTP_URL, FS_DOTFILES_PATH])
-    logging.info('Dotfiles downloaded---Downloading git prompt now')
+    logging.info('Dotfiles downloaded --- Downloading git prompt now')
     subprocess.call(['curl', GIT_COMPLETION_URL, '-o', os.path.join(HOME_DIR, '.git-prompt.sh')])
     logging.info('git-prompt.sh downloaded')
 
@@ -86,24 +102,32 @@ def main(args):
         util.add_stdout_logging()
     else:
         util.add_file_logging()
+
     download_dotfiles()
-    if args.d or not args.a:
+
+    if args.d or args.a:
         setup_dotfiles()
-    if args.e or not args.a:
+    if args.e or args.a:
         setup_emacs()
-    if args.t or not args.a:
+    if args.t or args.a:
         setup_tmux()
-    if args.g or not args.a:
+    if args.g or args.a:
         setup_git()
+    if args.v or args.a:
+        setup_vim()
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Bootstrap dotfiles')
     parser.add_argument('--l', action='store_false', default=True, help='Dump logging to stdout/stderr')
-    parser.add_argument('--a', action='store_true', default=False, help='Don\'t setup all')
-    parser.add_argument('--d', action='store_true', default=False, help='Explicit setup dotfiles')
-    parser.add_argument('--e', action='store_true', default=False, help='Explicit setup emacs')
-    parser.add_argument('--t', action='store_true', default=False, help='Explicit setup tmux')
-    parser.add_argument('--g', action='store_true', default=False, help='Explicit setup git')
+
+    parser.add_argument('--a', action='store_true', default=False, help='Setup all')
+
+    parser.add_argument('--d', action='store_true', default=False, help='Setup dotfiles')
+    parser.add_argument('--e', action='store_true', default=False, help='Setup emacs')
+    parser.add_argument('--v', action='store_true', default=False, help='Setup vim')
+    parser.add_argument('--t', action='store_true', default=False, help='Setup tmux')
+    parser.add_argument('--g', action='store_true', default=False, help='Setup git')
+
     args = parser.parse_args()
 
     main(args)
